refactor: rename components and pages
